Revert "Move joinRight() and joinBottom() from Matrix to AbstractMatrix"
This reverts commit c9a3ab9c9cdeefa09ff5743a03d7cac336c13d8a.
