+++ src/LinearAlgebra/Matrix.php
@@ -753,4 +753,82 @@ class Matrix extends AbstractMatrix
 
         return $determinant;
     }
+
+    /**
+     * Horizontal matrix concatenation (augmentation from the right)
+     *
+     * @param Matrix $anotherMatrix
+     * @return static
+     * @throws InvalidArgumentException (not expected)
+     * @throws MatrixException if the matrices have different amount of rows
+     */
+    public function joinRight(Matrix $anotherMatrix): static
+    {
+        $newMatrix = new static($this->matrix, false);
+        return $newMatrix->mJoinRight($anotherMatrix);
+    }
+
+    /**
+     * Horizontal matrix concatenation (augmentation from the right). Mutating method changes the original matrix
+     *
+     * @param Matrix $anotherMatrix
+     * @return $this
+     * @throws MatrixException if the matrices have different amount of rows
+     * @internal Mutating method
+     */
+    public function mJoinRight(Matrix $anotherMatrix): static
+    {
+        // Check if the matrices are compatible
+        if ($this->rows !== $anotherMatrix->rows) {
+            throw new MatrixException("Cannot perform horizontal concatenation. Both matrices must have same number of rows.");
+        }
+
+        foreach ($this->matrix as $rowIndex => &$row) {
+            $row = array_merge($row, $anotherMatrix->matrix[$rowIndex]);
+        }
+
+        // Clear cache and recalculate matrix dimensions
+        $this->clearCache();
+        $this->columns += $anotherMatrix->columns;
+        $this->size = $this->rows * $this->columns;
+        return $this;
+    }
+
+    /**
+     * Vertical matrix concatenation (augmentation from the bottom)
+     *
+     * @param Matrix $anotherMatrix
+     * @return static
+     * @throws InvalidArgumentException (not expected)
+     * @throws MatrixException if the matrices have different amount of columns
+     */
+    public function joinBottom(Matrix $anotherMatrix): static
+    {
+        $newMatrix = new static($this->matrix, false);
+        return $newMatrix->mJoinBottom($anotherMatrix);
+    }
+
+    /**
+     * Vertical matrix concatenation (augmentation from the bottom). Mutating method changes the original matrix
+     *
+     * @param Matrix $anotherMatrix
+     * @return $this
+     * @throws MatrixException if the matrices have different amount of columns
+     * @internal Mutating method
+     */
+    public function mJoinBottom(Matrix $anotherMatrix): static
+    {
+        // Check if the matrices are compatible
+        if ($this->columns !== $anotherMatrix->columns) {
+            throw new MatrixException("Cannot perform vertical concatenation. Both matrices must have same number of columns.");
+        }
+
+        $this->matrix = array_merge($this->matrix, $anotherMatrix->matrix);
+
+        // Clear cache and recalculate matrix dimensions
+        $this->clearCache();
+        $this->rows += $anotherMatrix->rows;
+        $this->size = $this->rows * $this->columns;
+        return $this;
+    }
 }
