+++ src/LinearAlgebra/Vector.php
@@ -329,5 +329,4 @@ class Vector extends Matrix
 
         return $dotProduct;
     }
-
 }
